# swagger.yaml
openapi: 3.0.0
info:
  title: Library API
  version: 1.0.0
  description: Capstone Backend Project - Library Management System

servers:
  - url: http://localhost:3000
    description: Local Development

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: "John Doe" }
        email: { type: string, example: "john@email.com" }
        password: { type: string, example: "password123" }
    
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: "john@email.com" }
        password: { type: string, example: "password123" }
    
    AuthResponse:
      type: object
      properties:
        token: { type: string, example: "eyJhbGciOiJIUzI1NiIs..." }
    
    Book:
      type: object
      properties:
        id: { type: integer, example: 1 }
        title: { type: string, example: "Atomic Habits" }
        author: { type: string, example: "James Clear" }
        stock: { type: integer, example: 5 }
    
    Borrow:
      type: object
      properties:
        id: { type: integer, example: 1 }
        user_name: { type: string, example: "John Doe" }
        book_title: { type: string, example: "Atomic Habits" }
        borrow_date: { type: string, format: date, example: "2026-06-20" }
        return_date: { type: string, format: date, example: null }
        status: { type: string, example: "Borrowed" }

paths:
  /auth/register:
    post:
      summary: Register User
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: User registered successfully
        400:
          description: Validation error
        500:
          description: Server error

  /auth/login:
    post:
      summary: Login User
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Invalid credentials
        404:
          description: User not found

  /books:
    get:
      summary: Get All Books
      tags: [Books]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        401:
          description: Unauthorized

    post:
      summary: Create Book
      tags: [Books]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, author, stock]
              properties:
                title: { type: string, example: "New Book" }
                author: { type: string, example: "Author Name" }
                stock: { type: integer, example: 10 }
      responses:
        201:
          description: Book created
        400:
          description: Validation error
        401:
          description: Unauthorized

  /books/{id}:
    get:
      summary: Get Book by ID
      tags: [Books]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Book found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        404:
          description: Book not found

    put:
      summary: Update Book
      tags: [Books]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                author: { type: string }
                stock: { type: integer }
      responses:
        200:
          description: Book updated
        400:
          description: Validation error
        401:
          description: Unauthorized

    delete:
      summary: Delete Book
      tags: [Books]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Book deleted
        401:
          description: Unauthorized

  /borrow:
    get:
      summary: Get All Borrowings
      tags: [Borrow]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of borrowings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Borrow'
        401:
          description: Unauthorized

    post:
      summary: Borrow a Book
      tags: [Borrow]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [book_id]
              properties:
                book_id: { type: integer, example: 1 }
      responses:
        201:
          description: Book borrowed
        400:
          description: Invalid request or out of stock
        401:
          description: Unauthorized
        404:
          description: Book not found

  /borrow/return/{id}:
    put:
      summary: Return a Book
      tags: [Borrow]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Book returned
        401:
          description: Unauthorized
        404:
          description: Borrow record not found